# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Franklins AI Driver Onboarding - Development/Demo API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drivers:
    get:
      operationId: listDrivers
      tags: [drivers]
      summary: List all onboarding drivers
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: driverType
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: assigneeId
          in: query
          schema:
            type: ["number", "null"]
      responses:
        "200":
          description: List of drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
    post:
      operationId: createDriver
      tags: [drivers]
      summary: Create a new onboarding record (simulate Hired event)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "409":
          description: Duplicate - driver already exists for this recruit event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /drivers/operational-queue:
    get:
      operationId: listDriverOperationalQueue
      tags: [drivers]
      summary: List workspace-scoped operational driver queue
      parameters:
        - name: view
          in: query
          schema:
            type: string
            enum: [all, due_today, overdue, waiting_blocked, no_next_action, needs_review]
        - name: status
          in: query
          schema: { type: string }
        - name: driverType
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: operationalOwner
          in: query
          schema: { type: string }
        - name: step
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 11
        - name: source
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Filtered operational queue and counts from the same dataset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverOperationalQueue"
  /drivers/{id}:
    get:
      operationId: getDriver
      tags: [drivers]
      summary: Get a single driver onboarding record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Driver record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateDriver
      tags: [drivers]
      summary: Update a driver onboarding record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /drivers/{id}/datatruck-sync:
    post:
      operationId: triggerDatatruckSync
      tags: [drivers]
      summary: Simulate idempotent DataTruck TMS sync
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Sync record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatatruckSync"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /drivers/{id}/datatruck-sync/history:
    get:
      operationId: getDatatruckSyncHistory
      tags: [drivers]
      summary: Get DataTruck sync audit history for a driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of sync records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DatatruckSync"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean
      required:
        - code
        - message
        - retryable
    Driver:
      type: object
      required:
        - id
        - fullName
        - driverType
        - status
        - stage
        - recruiterName
        - sourceChannel
        - createdAt
        - updatedAt
      properties:
        id:
          type: number
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        driverType:
          type: string
          enum: [owner_operator, company_driver]
        status:
          type: string
          enum: [pre_hire, in_progress, pending_approval, approved, ready_for_dispatch, dispatched, fallout, on_hold]
        stage:
          type: string
        priority:
          type: string
          enum: [critical, high, medium, low]
        recruiterName:
          type: string
        sourceChannel:
          type: string
        assigneeId:
          type: ["number", "null"]
        assigneeName:
          type: ["string", "null"]
        operationalOwnerId:
          type: ["number", "null"]
        operationalOwnerName:
          type: ["string", "null"]
        truckVin:
          type: ["string", "null"]
        truckInfo:
          type: ["string", "null"]
        telegramGroupLinked:
          type: boolean
        readyForDispatch:
          type: boolean
        datatruckSyncStatus:
          type: ["string", "null"]
          enum: [pending, synced, failed, retry, "null"]
        startDate:
          type: ["string", "null"]
        slaDeadline:
          type: ["string", "null"]
        blockers:
          type: ["string", "null"]
        nextBestAction:
          type: ["string", "null"]
        externalRecruitId:
          type: ["string", "null"]
        complianceGatesPassed:
          type: boolean
        completionPercent:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
...
    DriverOperationalQueue:
      type: object
      required: [items, counts, total, stepCount, filterOptions]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/DriverOperationalRow" }
        counts:
          type: object
          required: [all, dueToday, overdue, waitingBlocked, noNextAction, needsReview]
          properties:
            all: { type: number }
            dueToday: { type: number }
            overdue: { type: number }
            waitingBlocked: { type: number }
            noNextAction: { type: number }
            needsReview: { type: number }
        total: { type: number }
        stepCount:
          type: number
          description: Number of canonical operational milestones in this workspace workflow.
        filterOptions:
          type: object
          required: [owners, steps, sources]
          properties:
            owners:
              type: array
              items:
                type: object
                required: [id, name]
                properties:
                  id: { type: number }
                  name: { type: string }
            steps:
              type: array
              items:
                type: object
                required: [number, key, label]
                properties:
                  number: { type: number }
                  key: { type: string }
                  label: { type: string }
            sources:
              type: array
              items: { type: string }
    DriverInput:
      type: object
      required:
        - fullName
        - driverType
        - recruiterName
        - sourceChannel
      properties:
        fullName:
          type: string
          minLength: 1
        phone:
          type: string
        email:
          type: string
        state:
          type: string
        driverType:
          type: string
          enum: [owner_operator, company_driver]
        recruiterName:
          type: string
        sourceChannel:
          type: string
        assigneeId:
          type: number
        truckVin:
          type: string
        startDate:
          type: string
        externalRecruitId:
          type: string
        priority:
          type: string
          enum: [critical, high, medium, low]
    DriverUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        state:
          type: string
        status:
          type: string
        stage:
          type: string
        priority:
          type: string
        assigneeId:
          type: ["number", "null"]
        truckVin:
          type: string
        truckInfo:
          type: string
        telegramGroupLinked:
          type: boolean
        startDate:
          type: string
        slaDeadline:
          type: string
        blockers:
          type: string
        nextBestAction:
          type: string
...
    DatatruckSync:
      type: object
      required:
        - id
        - driverId
        - syncStatus
        - createdAt
      properties:
        id:
          type: number
        driverId:
          type: number
        syncStatus:
          type: string
          enum: [pending, synced, failed, retry]
        attemptNumber:
          type: number
        errorMessage:
          type: ["string", "null"]
        syncedAt:
          type: ["string", "null"]
        createdAt:
          type: string
...